Fake-New-Detection-With-LSTM-And-Glove
In the NLP project I did to detect fake news, I used the LSTM artificial neural network with the 300-dimensional GloVe word embedding vector. Here, I used the GloVe model 300 because the texts in my data set were very long, and for the same reason, choosing the LSTM artificial neural network played a big role in the accuracy of the model.